refactor: move fn read_test_resource into test_utils.rs
